requirements.txt
agate==1.14.2
agate-dbf==0.2.4
agate-excel==0.4.2
agate-sql==0.7.3
aio-pika==9.5.8
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiokafka==0.12.0
aiormq==6.9.2
aiosignal==1.4.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asgiref==3.11.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
asyncpg==0.31.0
asyncssh==2.21.1
attrs==25.4.0
autopep8==2.3.2
azure-core==1.37.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
azure-kusto-data==6.0.0
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
billiard==4.2.4
bleach==6.3.0
blinker==1.9.0
boto3==1.42.9
botocore==1.42.9
cachetools==6.2.3
celery==5.6.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
csvkit==2.2.0
cycler==0.12.1
dbfread==2.0.7
debugpy==1.8.18
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docopt==0.6.2
duckdb==1.4.3
et_xmlfile==2.0.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.128.0
fastapi-cache2==0.2.2
fastapi-limiter==0.1.6
fastjsonschema==2.21.2
fastuuid==0.14.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.2
fonttools==4.61.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.4
google-genai==1.60.0
googleapis-common-protos==1.72.0
grpcio==1.67.1
grpcio-status==1.67.1
gspread==6.2.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hiredis==3.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
hyperframe==6.1.0
idna==3.11
ijson==3.4.0.post0
importlib_metadata==8.7.0
invoke==2.2.1
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jira==3.10.5
jiter==0.12.0
jmespath==1.0.1
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.1
lark==1.3.1
leather==0.4.1
litellm==1.80.9
lxml==6.0.2
mail-parser==4.1.4
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mimesis==18.0.0
mistune==3.1.4
motor==3.7.1
msal==1.34.0
msal-extensions==1.3.1
multidict==6.7.0
mypy_extensions==1.1.0
namedlist==1.8
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.5.0
notebook_shim==0.2.4
numpy==2.3.5
oauthlib==3.3.1
olefile==0.47
ollama==0.6.1
openai==2.11.0
openpyxl==3.1.5
orjson==3.11.6
overrides==7.7.0
packaging==25.0
pamqp==3.3.0
pandas==2.3.3
pandocfilters==1.5.1
paramiko==4.0.0
parsedatetime==2.6
parso==0.8.5
pathspec==0.12.1
pdf2image==1.17.0
pendulum==3.1.0
pexpect==4.9.0
pika==1.3.2
pillow==12.0.0
pipreqs==0.4.13
platformdirs==4.5.1
portalocker==3.2.0
posthog==7.0.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
psycopg==3.3.3
psycopg-binary==3.3.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.5
PyMuPDF==1.26.7
PyNaCl==1.6.1
pyodbc==5.3.0
pyparsing==3.2.5
pypdf==6.4.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
Send2Trash==1.8.3
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
stack-data==0.6.3
starlette==0.50.0
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
tiktoken==0.12.0
time-machine==3.1.0
tinycss2==1.4.0
tokenizers==0.22.1
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
types-python-dateutil==2.9.0.20251115
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
vega-datasets==0.9.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.4
widgetsnbextension==4.0.15
xlrd==2.0.2
yarg==0.1.10
yarl==1.22.0
zipp==3.23.0
